% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestart.R
\name{prestart}
\alias{prestart}
\title{Prestart package \ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}}
\usage{
prestart()
}
\value{

}
\description{
Prestart package \ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
