% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestart-package.R
\docType{package}
\name{prestart-package}
\alias{prestart}
\alias{prestart-package}
\title{prestart: Set up bare minimum for tidy package dev}
\description{
Set up bare minimum for tidy package dev.
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko.thyson@rappster.io}

}
\keyword{internal}
